Add Programmatic & Bean-Injection use of SpEL expression language. (#19)
SpEL was introduced in Spring 3.0 in Dec 2009, so many years!

This part I add an `Evaluator` class which is a programmatic way on how to use SpEL to parse expressions.

Also add an `EvaluatorComponent` bean, demonstrate in Spring container how we use SpEL to inject values.

There are much more SpEL powerful capabilities need further investigating.

core concept in SpEL is ast. (a new lesson for me :) )

more see:

https://docs.spring.io/spring-framework/reference/core/expressions.html

https://youtu.be/0uvQQuxyAv4

Signed-off-by: Jiandong Ma <jiandong.ma.cn@gmail.com>